# openapi.yaml
openapi: 3.1.0
info:
  title: BolusGPT API
  description: API for insulin dosing and user metabolic settings.
  version: 1.0.0
servers:
  - url: https://foo.bar
    description: Production server
paths:
  /me:
    get:
      operationId: getMe
      summary: Get user settings
      description: Returns the currently stored user metabolic configuration.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
              examples:
                currentConfig:
                  summary: Example user metabolic configuration
                  value:
                    fiber_multiplier: 0.5
                    sugar_alcohol_multiplier: 0.5
                    protein_multiplier: 0.1
                    carb_threshold_to_count_protein_under: 25
                    insulin_to_carb_ratio: 10
                    target_blood_glucose_level_in_mg_dl: 100
                    insulin_sensitivity_factor: 40
                    last_bolus_time: "2025-04-07T12:00:00Z"
                    last_bolus_units_of_insulin: 4.0
        '404':
          description: User has not onboarded
    patch:
      operationId: updateMe
      summary: Update user settings
      description: Updates user settings such as multipliers, sensitivity, and recent insulin usage. Returns the updated config.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeInput'
            examples:
              onboardUser:
                summary: Onboard a new user with full metabolic config
                value:
                  fiber_multiplier: 0.5
                  sugar_alcohol_multiplier: 0.25
                  protein_multiplier: 0.1
                  carb_threshold_to_count_protein_under: 20
                  insulin_to_carb_ratio: 12
                  insulin_sensitivity_factor: 45
                  target_blood_glucose_level_in_mg_dl: 110
              updateCarbRatio:
                summary: Set insulin-to-carb ratio to 5
                value:
                  insulin_to_carb_ratio: 5
              setFiberMultiplier:
                summary: Count 50% of fiber in carb calculation
                value:
                  fiber_multiplier: 0.5
              confirmBolus:
                summary: Log that 5 units were taken now
                value:
                  last_bolus_units_of_insulin: 5
                  last_bolus_time: "2025-04-07T15:30:00Z"
      responses:
        '200':
          description: Updated user configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '500':
          description: Server error
  /dose:
    post:
      operationId: getDose
      summary: Calculate insulin dose
      description: Calculates insulin dose based on input data and current user settings.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoseInput'
            examples:
              yogurtOnly:
                summary: Bolus dose for a cup of yogurt
                value:
                  total_grams_of_carbs: 15
                  grams_of_fiber: 0
                  grams_of_sugar_alcohol: 0
                  grams_of_protein: 5
              yogurtWithRun:
                summary: Bolus dose after a long, heavy run
                value:
                  total_grams_of_carbs: 15
                  grams_of_fiber: 0
                  grams_of_sugar_alcohol: 0
                  grams_of_protein: 5
                  minutes_of_exercise: 60
                  exercise_intensity: high
      responses:
        '200':
          description: Calculated dose
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dose'
        '404':
          description: Missing required configuration
        '500':
          description: Server error
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Me:
      type: object
      properties:
        fiber_multiplier:
          type: number
          description: Adjustment factor for dietary fiber's effect on insulin needs.
        sugar_alcohol_multiplier:
          type: number
          description: Adjustment factor for sugar alcohols' impact on blood sugar.
        protein_multiplier:
          type: number
          description: Factor representing how protein contributes to insulin demand.
        carb_threshold_to_count_protein_under:
          type: number
          description: Carb threshold under which protein is counted for dosing.
        insulin_to_carb_ratio:
          type: number
          description: Grams of carbs covered by one unit of insulin.
        target_blood_glucose_level_in_mg_dl:
          type: number
          description: Target blood glucose level in mg/dL.
        insulin_sensitivity_factor:
          type: number
          description: Blood glucose drop expected per unit of insulin.
        last_bolus_time:
          type: string
          format: date-time
          description: Time of the last insulin bolus.
        last_bolus_units_of_insulin:
          type: number
          description: Amount of insulin used in the last bolus.
    MeInput:
      type: object
      properties:
        fiber_multiplier:
          type: number
          description: Adjustment factor for dietary fiber's effect on insulin needs.
        sugar_alcohol_multiplier:
          type: number
          description: Adjustment factor for sugar alcohols' impact on blood sugar.
        protein_multiplier:
          type: number
          description: Factor representing how protein contributes to insulin demand.
        carb_threshold_to_count_protein_under:
          type: number
          description: Carb threshold under which protein is counted for dosing.
        insulin_to_carb_ratio:
          type: number
          description: Grams of carbs covered by one unit of insulin.
        target_blood_glucose_level_in_mg_dl:
          type: number
          description: Target blood glucose level in mg/dL.
        insulin_sensitivity_factor:
          type: number
          description: Blood glucose drop expected per unit of insulin.
        last_bolus_time:
          type: string
          format: date-time
          description: Time of the last insulin bolus.
        last_bolus_units_of_insulin:
          type: number
          description: Amount of insulin used in the last bolus.
    DoseInput:
      type: object
      properties:
        total_grams_of_carbs:
          type: number
          description: Total carbohydrates consumed.
        grams_of_fiber:
          type: number
          description: Grams of dietary fiber in the meal.
        grams_of_sugar_alcohol:
          type: number
          description: Grams of sugar alcohols in the meal.
        grams_of_protein:
          type: number
          description: Grams of protein in the meal.
        minutes_of_exercise:
          type: number
          description: Duration of recent exercise in minutes.
        exercise_intensity:
          type: string
          description: Intensity of recent exercise.
          enum: [none, low, medium, high]
    Dose:
      type: object
      description: Output from the dose calculation.
      properties:
        units_of_insulin:
          type: number
          description: Units of Insulin for the Bolus dose.
        grams_of_carbs:
          type: number
          description: If negative, carbs to consume to return to target blood glucose.
        breakdown:
          type: object
          description: A breakdown of factors contributing to the total dose.
          properties:
            food_factor:
              type: number
              description: Portion of dose due to food intake.
            correction_factor:
              type: number
              description: Portion of dose for correcting high blood glucose.
            insulin_on_board_factor:
              type: number
              description: Portion of dose adjusted for insulin still active in the body.
            exercise_multiplier:
              type: number
              description: Portion of dose adjusted due to recent exercise.